r1629: server_zone needs to be "int" not "int16_t" as it can hold values larger than 35536 (the max value is actually 12*3600 == 43200) (This used to be commit 78dcaddbb78574b7f489989df0b1f979677bb7d5)

+++ source4/include/cli_context.h
@@ -63,7 +63,7 @@ struct cli_negotiate {
 	/* capabilities that the server reported */
 	uint32_t capabilities;
 	
-	int16_t server_zone;
+	int server_zone;
 	time_t server_time;
 	uint_t readbraw_supported:1;
 	uint_t writebraw_supported:1;
